added use of ensureMemoryLimit for invoices.
git-svn-id: https://dev.virtuemart.net/svn/virtuemart/branches/com_virtuemart2.0.14mp@7167 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ administrator/components/com_virtuemart/helpers/config.php
@@ -463,7 +463,7 @@ class VmConfig {
 	 */
 	static function ensureMemoryLimit($minMemory=0){
 
-		if($minMemory === 0) $minMemory = (int) VmConfig::get('minMemory','128M');
+		if($minMemory === 0) $minMemory = VmConfig::get('minMemory','128M');
 		$iniValue = ini_get('memory_limit');
 		if($iniValue===-1) return;	//We do not need to alter an unlimited setting
 		$iniValue = strtolower($iniValue);
